+++ src/features/profile/api/getProfile.ts
@@ -1,0 +1,6 @@
+import { API_END_POINT, httpClient } from "@/shared";
+
+export const getProfile = async () => {
+  const response = await httpClient.get(`${API_END_POINT.SIGNUP}`);
+  return response.data;
+};
