+++ src/features/profile/ui/index.ts
@@ -1,0 +1,2 @@
+export { ProfileImageUploader } from './ProfileImageUploader';
+export { ProfileInput } from './ProfileInput';
