+++ src/pages/user/ui/UserProfileEdit.tsx
@@ -5,14 +5,14 @@ import { useDispatch, useSelector } from 'react-redux';
 import { IUserProfile } from '@/@types/user';
 import { Layout } from '@/app/layout/index';
 import { RootState } from '@/app/store';
-import ProfileImageUploader from '@/components/profile/ProfileImageUploader';
-import { useCheckNickname } from '@/components/profile/queries';
 import { useEditProfile } from '@/hooks/useProfile';
 import { Button, FormField } from '@/shared';
 import NoticeBlue from '@/shared/assets/icons/blue_notice.svg';
 import NoticeRed from '@/shared/assets/icons/notice_red.svg';
 import { Input } from '@/shared/shadcn/ui/input';
 import { Textarea } from '@/shared/shadcn/ui/textarea';
+import { ProfileImageUploader } from '@/features/profile/ui';
+import { useCheckNickname } from '@/features/profile/model/useProfile';
 
 export const UserProfileEdit = () => {
   const formRef = useRef<HTMLFormElement>(null);
@@ -117,7 +117,7 @@ export const UserProfileEdit = () => {
           onSubmit={handleSubmit(onSubmit)}
         >
           <h2 className="pb-4 text-heading3 web:text-heading2">프로필 정보</h2>
-          <ProfileImageUploader 
+          <ProfileImageUploader
             file={profileFile}
             setFile={setProfileFile}
             image={profileImage}
